% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chisquare_test.R
\name{Chisquare_test}
\alias{Chisquare_test}
\title{Chi-square statistics}
\usage{
Chisquare_test(x, nrow, ncol)
}
\arguments{
\item{x}{Data matrix X}

\item{nrow}{Number of rows of data matrix X}

\item{ncol}{Number of columns of data matrix X}
}
\description{
Chi-square statistics
}
