% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportion_test.R
\name{f1}
\alias{f1}
\title{Hypergeometric distribution interval estimation}
\usage{
f1(n, N, k, t0, p)
}
\arguments{
\item{n}{Sample size}

\item{N}{Total number}

\item{k}{k=N*theta}

\item{t0}{Number of data with specific attributes}

\item{p}{P value}
}
\value{
p
}
\description{
Hypergeometric distribution interval estimation
}
