% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportion_test.R
\name{f2}
\alias{f2}
\title{Estimation of Hypergeometric Distribution Interval under Large Sample}
\usage{
f2(p, n, t)
}
\arguments{
\item{p}{theta}

\item{n}{Sample size}

\item{t}{Number of data with specific attributes}
}
\value{
s
}
\description{
Estimation of Hypergeometric Distribution Interval under Large Sample
}
