% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportion_test.R
\name{find_k1k2_1}
\alias{find_k1k2_1}
\title{Hypergeometric distribution interval estimation}
\usage{
find_k1k2_1(n, N, k, a)
}
\arguments{
\item{n}{Sample size}

\item{N}{Total number}

\item{k}{k=N*theta}

\item{a}{Significance level}
}
\value{
vec
}
\description{
Hypergeometric distribution interval estimation
}
