% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportion_test.R
\name{find_k1k2_2}
\alias{find_k1k2_2}
\title{Estimation of Hypergeometric Distribution Interval under Large Sample}
\usage{
find_k1k2_2(p, n, a)
}
\arguments{
\item{p}{P value}

\item{n}{Sample size}

\item{a}{Significance level}
}
\value{
vec
}
\description{
Estimation of Hypergeometric Distribution Interval under Large Sample
}
