\name{StatComp20068-package}
\alias{StatComp20068-package}
\alias{StatComp20068}
\docType{package}
\title{
   Chi-square statistics & The test of overall proportion
}
\description{
  The software package provides the population proportion test function and the chi-square statistic calculation function that obey the hypergeometric distribution.
}

\author{
Qiuyang Li

Maintainer: Qiuyang Li <2213847080@qq.com>
}
\keyword{ package }

